Show users with bad API keys in the guild page, and allow filtering guild roster snapshots by raid teams
